Patch the pkg-config directory in the gz-cmake code. (#4)
* Patch the pkg-config directory in the gz-cmake code.

When building on the ROS 2 buildfarm, we aren't setting
some of the CMAKE_PREFIX variables.  This means that
using CMAKE_INSTALL_FULL_LIBDIR actually creates a path
like /opt/ros/rolling/... , which makes debuild upset.

However, we actually need the FULL_LIBDIR in order to
calculate the relative path between it and the INSTALL_PREFIX.
Work around this by having two variables; the
pkgconfig_install_dir (relative), used to install the files,
and pkgconfig_abs_install_dir (absolute), used to calculate
the relative path between them.

This should fix the build on the buildfarm.  I'll note that
we are doing it here and not in gz-cmake proper because of
knock-on effects to downstream gazebo.  If this is successful
we may end up merging it there, at which point we can drop
this patch.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>

* Update GzPackage as well.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>

---------

Signed-off-by: Chris Lalancette <clalancette@gmail.com>